+++ audionimbus/CHANGELOG.md
@@ -18,6 +18,8 @@
 - Fixed a segmentation fault where `ReflectionEffect::apply_into_mixer` was missing the output buffer.
 - Fixed a segmentation fault when applying a binaural effect on an input buffer other than mono or stereo, or passing an output buffer that has more than two channels.
 - Fixed a segmentation fault when calling `BinauralEffect::tail` with an output buffer that has more than two channels.
+- Fixed a segmentation fault when applying a panning effect on an input buffer other than mono, or passing an output buffer that has a number of channels different from that needed for the speaker layout specified when creating the effect.
+- Fixed a segmentation fault when calling `BinauralEffect::tail` with an output buffer that has a number of channels different from that needed for the speaker layout specified when creating the effect.
 
 ### Changed
 
@@ -39,7 +41,9 @@
 - `ReflectionEffect::apply` and `ReflectionEffect::apply_into_mixer` now return an `EffectError` when the input buffer is not mono or the output buffer has a number of channels other than that of the impulse response specified when creating the effect.
 - `ReflectionEffect::tail` and `ReflectionEffect::tail_into_mixer` now return an `EffectError` when the output buffer has a number of channels other than that of the impulse response specified when creating the effect.
 - `BinauralEffect::apply` now returns an `EffectError` when the input buffer is not mono or stereo, or the output buffer has more than two channels.
-- `BinauralEffect::tail` now returns an `EffectError` when the output buffer has more than two channels
+- `BinauralEffect::tail` now returns an `EffectError` when the output buffer has more than two channels.
+- `PanningEffect::apply` now returns an `EffectError` when the input buffer is not mono, or the output buffer has a number of channels different from that needed for the speaker layout specified when creating the effect.
+- `PanningEffect::tail` now returns an `EffectError` when the output buffer has a number of channels different from that needed for the speaker layout specified when creating the effect.
 
 ### Added